DF: escapeMenu: Fix wrong palette in german version.
Apply the "menu.PLTT" palette to the escape Menu.
Fixes the off colors in the German version of the quit screen.

+++ TheForceEngine/TFE_DarkForces/mission.h
@@ -47,5 +47,5 @@ namespace TFE_DarkForces
 	extern u8 s_loadingScreenPal[];
 	extern u8 s_levelPalette[];
 	extern u8 s_basePalette[];
-	extern u8 s_escMenuPalette[];
+	extern u32 s_escMenuPalette[];
 }  // namespace TFE_DarkForces
